feat: 결제 승인 구현 완료

+++ src/main/java/com/sparta/nugulpayment/payment/dto/request/PreprocessRequest.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class PreprocessRequest {
     private String orderId;
-    private int amount;
+    private long amount;
     private long userId;
 }
